requirements.txt
Django==4.2.7
psycopg2-binary
djangorestframework
channels
channels-redis